# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Schedule Manager API
  version: 1.0.0
  description: React + Firebase Auth + Express + Firestore 일정관리 API

servers:
  - url: http://localhost:3000
    description: Local server

tags:
  - name: Schedules
    description: 일정 관리 API
  - name: Attendance
    description: 근태 관리 API

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token을 Bearer 토큰으로 입력하세요.

  schemas:
    Attendance:
      type: object
      properties:
        id:
          type: string
          example: att_001
        userId:
          type: string
          example: firebase_uid_123
        workDate:
          type: string
          example: "2026-04-16"
        checkInAt:
          nullable: true
          oneOf:
            - type: string
              format: date-time
              example: "2026-04-16T08:57:12.000Z"
            - type: "null"
        checkOutAt:
          nullable: true
          oneOf:
            - type: string
              format: date-time
              example: "2026-04-16T18:05:10.000Z"
            - type: "null"
        status:
          type: string
          example: present
        totalWorkMinutes:
          type: integer
          example: 548
        note:
          type: string
          example: ""
        createdAt:
          type: string
          format: date-time
          example: "2026-04-16T08:57:12.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-16T18:05:10.000Z"
    Schedule:
      type: object
      properties:
        id:
          type: string
          example: abc123xyz
        title:
          type: string
          example: 팀 미팅
        start:
          type: string
          format: date-time
          example: 2026-04-16T09:00:00.000Z
        end:
          type: string
          format: date-time
          example: 2026-04-16T10:00:00.000Z
        memo:
          type: string
          example: 주간 회의
        color:
          type: string
          example: "#3b82f6"
        userId:
          type: string
          example: firebase_uid_123
        createdAt:
          type: string
          format: date-time
          example: 2026-04-16T01:00:00.000Z
        updatedAt:
          nullable: true
          oneOf:
            - type: string
              format: date-time
              example: 2026-04-16T02:00:00.000Z
            - type: "null"

    ScheduleCreateRequest:
      type: object
      required:
        - title
        - start
        - end
      properties:
        title:
          type: string
          example: Swagger 테스트 일정
        start:
          type: string
          format: date-time
          example: 2026-04-16T09:00:00.000Z
        end:
          type: string
          format: date-time
          example: 2026-04-16T10:00:00.000Z
        memo:
          type: string
          example: 스웨거에서 생성한 일정
        color:
          type: string
          example: "#3b82f6"

    ScheduleUpdateRequest:
      type: object
      properties:
        title:
          type: string
          example: 수정된 일정 제목
        start:
          type: string
          format: date-time
          example: 2026-04-16T11:00:00.000Z
        end:
          type: string
          format: date-time
          example: 2026-04-16T12:00:00.000Z
        memo:
          type: string
          example: 수정된 메모
        color:
          type: string
          example: "#8b5cf6"

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: 일정이 삭제되었습니다.

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: 유효하지 않은 토큰입니다.

security:
  - bearerAuth: []

paths:
  /api/attendance/today:
    get:
      tags:
        - Attendance
      summary: 오늘 근태 조회
      description: 로그인한 사용자의 오늘 근태 기록을 조회합니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 오늘 근태 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasRecord:
                    type: boolean
                    example: true
                  workDate:
                    type: string
                    example: "2026-04-16"
                  record:
                    nullable: true
                    oneOf:
                      - $ref: "#/components/schemas/Attendance"
                      - type: "null"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/attendance/check-in:
    post:
      tags:
        - Attendance
      summary: 출근 처리
      description: 로그인한 사용자의 오늘 출근 기록을 생성합니다.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: 출근 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: 이미 출근했거나 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/attendance/check-out:
    post:
      tags:
        - Attendance
      summary: 퇴근 처리
      description: 로그인한 사용자의 오늘 퇴근 기록을 저장합니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 퇴근 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: 출근 기록 없음 또는 이미 퇴근함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /health:
    get:
      summary: 서버 상태 확인
      description: 서버가 정상 실행 중인지 확인합니다.
      responses:
        "200":
          description: 정상 응답
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true

  /api/schedules:
    get:
      tags:
        - Schedules
      summary: 전체 일정 조회
      description: 로그인한 사용자의 전체 일정을 조회합니다.
      responses:
        "200":
          description: 일정 목록 조회 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류

    post:
      tags:
        - Schedules
      summary: 일정 생성
      description: 새 일정을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleCreateRequest"
      responses:
        "201":
          description: 일정 생성 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 서버 오류

  /api/schedules/{id}:
    get:
      tags:
        - Schedules
      summary: 일정 단건 조회
      description: 특정 일정 1건을 조회합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 일정 ID
          schema:
            type: string
      responses:
        "200":
          description: 일정 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "401":
          description: 인증 실패
        "403":
          description: 권한 없음
        "404":
          description: 일정 없음
        "500":
          description: 서버 오류

    put:
      tags:
        - Schedules
      summary: 일정 수정
      description: 기존 일정을 수정합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 일정 ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdateRequest"
      responses:
        "200":
          description: 일정 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: 잘못된 요청
        "401":
          description: 인증 실패
        "403":
          description: 권한 없음
        "404":
          description: 일정 없음
        "500":
          description: 서버 오류

    delete:
      tags:
        - Schedules
      summary: 일정 삭제
      description: 특정 일정을 삭제합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 일정 ID
          schema:
            type: string
      responses:
        "200":
          description: 일정 삭제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: 인증 실패
        "403":
          description: 권한 없음
        "404":
          description: 일정 없음
        "500":
          description: 서버 오류